BUG/MEDIUM: servers: Don't try to reuse connection if we switched server.
In connect_server(), don't attempt to reuse the old connection if it's
targetting a different server than the one we're supposed to access, or
we will never be able to connect to a server if the first one we tried failed.

This should be backported to 1.9.

(cherry picked from commit bb3dac37a291047e4da17a67bc1d23cbf92df2aa)
Signed-off-by: Willy Tarreau <w@1wt.eu>