BUG/MEDIUM: servers: Flag the stream_interface on handshake error.
When creating a new outgoing connection, if we're using ALPN and waiting
for the handshake completion to choose the mux, and for some reason the
handshake failed, add the SI_FL_ERR flag to the stream_interface, so that
process_streams() knows the connection failed, and can attempt to retry,
instead of just hanging.

This should be backported to 1.9.

(cherry picked from commit 4f41751ad21d89ca1e93e6fd6dca209f0cc66573)
Signed-off-by: Willy Tarreau <w@1wt.eu>