BUG/MEDIUM: servers: Fail if we fail to allocate a conn_stream.
If, for some reason we failed to allocate a conn_stream when reusing an
existing connection, set srv_conn to NULL, so that we fail later, instead
of pretending all is right. This ends up giving a stream_interface with
no endpoint, and so the stream will never end.

This should be backported to 1.9.

(cherry picked from commit 134a2045bbae1de2dd267928367a1445e966013d)
Signed-off-by: Willy Tarreau <w@1wt.eu>