MEDIUM: servers: Be smarter when switching connections.
When connecting to a server, and reusing a connection, always attempt to give
the owner of the previous session one of its own connections, so that one
session won't be responsible for too many connections.

This should be backported to 1.9.

(cherry picked from commit c685d700fd233b7ddd2737039af0cc50b94df3b5)
Signed-off-by: Willy Tarreau <w@1wt.eu>