MINOR: channel: Add the function channel_add_input
This function must be called when new incoming data are pushed in the channel's
buffer. It updates the channel state and take care of the fast forwarding by
consuming right amount of data and decrementing "->to_forward" accordingly when
necessary. In fact, this patch just moves a part of ci_putblk in a dedicated
function.

This patch must be backported to 1.9.

(cherry picked from commit e64582929fd1171ebe4c84488a1fcf44a5fb6fbc)
Signed-off-by: Willy Tarreau <w@1wt.eu>