MINOR: stats/htx: Call channel_add_input instead of updating channel state by hand
This way we are sure the channel state is always correctly upadated, especially
the amount of data directly forwarded. For the stats applet, it is not a bug
because the fast forwarding is never used (the response is chunked and the HTX
extra field is always set to 0).

This patch must be backported to 1.9.

(cherry picked from commit 5adbeeb336d244176b748d4f1ec271584c32131b)
Signed-off-by: Willy Tarreau <w@1wt.eu>