BUG/MEDIUM: proto-htx: Set SI_FL_NOHALF on server side when request is done
In the function htx_end_request, the flag SI_FL_NOHALF must be set on the server
side once the request is in the state HTTP_MSG_DONE. But the response state was
checked before and the flag was only set when the response was also in the state
HTTP_MSG_DONE. Of course, it is not desirable.

This patch must be backported to 1.9.

(cherry picked from commit d01ce4003d31c14d1eaed50f6c0bd9a83c6296a2)
Signed-off-by: Willy Tarreau <w@1wt.eu>