MEDIUM: mux-h2: make h2c_decode_headers() support recoverable errors
When a decoding error is recoverable, we should emit a stream error and
not a connection error. This patch does this by carefully checking the
connection state before deciding to send a connection error. If only the
stream is in error, an RST_STREAM is sent.

(cherry picked from commit 259192370f182f2b4597fe1aecb8857e988b7c0e)
Signed-off-by: Willy Tarreau <w@1wt.eu>