MEDIUM: mux-h2: pass trailers to H1 (legacy mode)
When forwarding an H2 request to an H1 server, if the request doesn't
have a content-length header field, it is chunked. In this case it is
possible to send trailers to the server, which is what this patch does.
If the transfer is performed without chunking, then the trailers are
silently discarded.

(cherry picked from commit e2b05ccff566a48253d5a29c2f93320488c469d2)
Signed-off-by: Willy Tarreau <w@1wt.eu>