BUG/MINOR: cache/htx: Be sure to count partial trailers
When a chunked object is served from the cache, If the trailers are not pushed
in the channel's buffer in one time, we still have to count them in the total
written bytes in the buffer.

This patch must be backported to 1.9.

(cherry picked from commit 74b41ba025c66bf4d4d2a5dfa320ea8376acb36e)
Signed-off-by: Willy Tarreau <w@1wt.eu>