BUG/MINOR: stats/htx: Respect the reserve when the stats page is dumped
As for the cache applet, this one must respect the reserve on HTX streams. This
patch is tagged as MINOR because it is unlikely to fully fill the channel's
buffer. Some tests are already done to not process almost full buffer.

This patch must be backported to 1.9.

(cherry picked from commit 69fc88c605e40d8401dde8e53c7a6f23901f6a70)
Signed-off-by: Willy Tarreau <w@1wt.eu>