BUG/MINOR: cache: Disable the cache if any compression filter precedes it
We need to check if any compression filter precedes the cache filter. This is
only possible when the compression is configured in the frontend while the cache
filter is configured on the backend (via a cache-store action or
explicitly). This case cannot be detected during HAProxy startup. So in such
cases, the cache is disabled.

The patch must be backported to 1.9.

(cherry picked from commit 839791af0dd77dabcb24f70943111cd5fcb8696b)
Signed-off-by: Willy Tarreau <w@1wt.eu>