MINOR: checks: Store the proxy in checks.
Instead of assuming we have a server, store the proxy directly in struct
check, and use it instead of s->server.
This should be a no-op for now, but will be useful later when we change
mail checks to avoid having a server.

This should be backported to 1.9.

(cherry picked from commit c98aa1f18252f855cffe280153e60e496f532b6e)
Signed-off-by: Willy Tarreau <w@1wt.eu>