BUG/MINOR: backend: BE_LB_LKUP_CHTREE is a value, not a bit
There are a few instances where the lookup algo is tested against
BE_LB_LKUP_CHTREE using a binary "AND" operation while this macro
is a value among a set, and not a bit. The test happens to work
because the value is exactly 4 and no bit overlaps with the other
possible values but this is a latent bug waiting for a new LB algo
to appear to strike. At the moment the only other algo sharing a bit
with it is the "first" algo which is never supported in the same code
places.

This fix should be backported to maintained versions for safety if it
passes easily, otherwise it's not important as it will not fix any
visible issue.

(cherry picked from commit 6c30be52da3d949a8dd6fb5e2de7319c031e656e)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 48147c424680b7e887fb176662d58d87baa16098)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>