BUG/MEDIUM: h1: Make sure we destroy an inactive connectin that did shutw.
In h1_process(), if we have no associated stream, and the connection got a
shutw, then destroy it, it is unusable and it may be our last chance to do
so.

This should be backported to 1.9.

(cherry picked from commit 32d75ed30024a8b4b75fdf1032f208d378e8d049)
Signed-off-by: Willy Tarreau <w@1wt.eu>